+++ pages/forgot-password/index.js
@@ -1,0 +1,9 @@
+import ForgotPassword from '../../src/components/Forms/ForgotPassword/ForgotPassword'
+import Layout from '../../src/components/layout'
+export default function index() {
+  return (
+    <Layout>
+      <ForgotPassword />
+    </Layout>
+  )
+}
