+++ pages/signup/index.js
@@ -1,0 +1,9 @@
+import SignUp from '../../src/components/Forms/SignUp/SignUp'
+import Layout from '../../src/components/layout'
+export default function index() {
+  return (
+    <Layout>
+      <SignUp />
+    </Layout>
+  )
+}
